pyproject.toml
[tool.poetry]
name = "arendt-bot"
version = "0.1.0"
description = "A LangChain-based project"
authors = ["Derek Koloditch <derek.koloditch@gmail.com>"]
readme = "README.md"
packages = [{include = "*", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
langchain = ">=0.3.15"
langchain-anthropic = ">=0.3.4"
anthropic = ">=0.45.0"
requests = "^2.32.3"
langchain-community = "^0.3.17"
pypdf = "^5.2.0"
langchain-openai = "^0.3.4"
faiss-cpu = "^1.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3